requirements.txt
biopython
nose
coverage
